Move /pipelines to pipelines.go.
